+++ src/point.rs
@@ -13,7 +13,7 @@ pub struct Point2D {
 }
 
 impl Point2D {
-    pub fn from_raw(x: i32, y: i32) -> Self {
+    pub fn from_raw(x: usize, y: usize) -> Self {
         Self::new(X::new(x), Y::new(y))
     }
